# Conflicts:
#	Codebase/WebsiteFrontend/src/components/Verwalter.jsx

+++ Codebase/DatabaseBackend/backend/src/main/java/com/backend/terminplanungsassitent/databaseClasses/LehrpersonRepository.java
@@ -2,6 +2,6 @@ package com.backend.terminplanungsassitent.databaseClasses;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LehrpersonRepository extends JpaRepository<Lehrperson, Long> {
+public interface LehrpersonRepository extends JpaRepository<Lehrperson, Integer> {
 
 }
